Added DataToScores function into imports.py and called/created assertions for it in tests.py
